test(server): decouple enum completion tests from spx
Move 17 enum completion test groups to plain XGo fixtures without loading
spx or embedded package data. Preserve their existing source scenarios,
cursor positions, and assertion conditions.

Use the isolated classfile framework to cover source kinds, cross-file
duplicate members, class field type inference, and document updates. Add
UTF-16 completion coverage.

Extend the spx resource return regression to check excess return values,
retaining coverage previously supplied by the enum completion tests.
